+++ tests/test_config.py
@@ -5,17 +5,20 @@ from cnxdb.config import discover_settings
 
 def test_success(mocker):
     common_url = 'postgresql:///common'
-    super_url = 'postgresql:///common'
+    readonly_url = 'postgresql:///readonly'
+    super_url = 'postgresql:///super'
 
     _patch = {
         'DB_URL': common_url,
-        'SUPER_URL': super_url,
+        'DB_READONLY_URL': readonly_url,
+        'DB_SUPER_URL': super_url,
     }
     mocker.patch.dict('os.environ', _patch, clear=True)
 
     settings = discover_settings()
 
     assert settings['db.common.url'] == common_url
+    assert settings['db.readonly.url'] == readonly_url
     assert settings['db.super.url'] == super_url
 
 
@@ -29,7 +32,7 @@ def test_required(mocker):
     assert 'DB_URL' in exc_info.value.args[0]
 
 
-def test_super_url_not_required(mocker):
+def test_other_urls_not_required(mocker):
     common_url = 'postgresql:///common'
 
     _patch = {
@@ -40,6 +43,7 @@ def test_super_url_not_required(mocker):
     settings = discover_settings()
 
     assert settings['db.common.url'] == common_url
+    assert settings['db.readonly.url'] == common_url
     assert settings['db.super.url'] == common_url
 
 
@@ -49,6 +53,7 @@ def test_with_existing_settings(mocker):
 
     _patch = {
         'DB_URL': other_url,
+        'DB_READONLY_URL': other_url,
         'DB_SUPER_URL': other_url,
     }
     mocker.patch.dict('os.environ', _patch, clear=True)
@@ -59,4 +64,5 @@ def test_with_existing_settings(mocker):
     settings = discover_settings(existing_settings)
 
     assert settings['db.common.url'] == common_url
+    assert settings['db.readonly.url'] == other_url
     assert settings['db.super.url'] == other_url